Start splitting out to standalone file
