+++ src/OpenLoco/CMakeLists.txt
@@ -287,6 +287,7 @@ set(OLOCO_SOURCES
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/CompanyFaceSelection.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/CompanyList.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/CompanyWindow.cpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/Blueprints.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/Common.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/ConstructionTab.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/OverheadTab.cpp"
@@ -685,6 +686,7 @@ set(OLOCO_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Window.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/WindowManager.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/WindowType.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/Blueprints.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/Construction/Construction.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/News/News.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/src/Ui/Windows/ToolbarTopCommon.h"
